requirements.txt
# Telegram Bot Framework
python-telegram-bot[job-queue]>=21.0

# Google Gemini AI
google-generativeai==0.8.5

# Database
SQLAlchemy==2.0.43
psycopg2-binary>=2.9.10  # Changed: Use latest version compatible with Python 3.13

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Environment Variables
python-dotenv==1.0.0

# Date/Time handling
python-dateutil>=2.8.2

# Async support
aiofiles>=23.2.1
pytz>=2023.3
boto3==1.34.162  # ✅ ADD THIS
